move save to project node
